Feature/sup 52/item

+++ src/main/java/by/itech/lab/supplier/service/impl/ItemServiceImpl.java
@@ -5,6 +5,7 @@ import by.itech.lab.supplier.dto.CategoryDto;
 import by.itech.lab.supplier.dto.ItemDto;
 import by.itech.lab.supplier.dto.mapper.ItemMapper;
 import by.itech.lab.supplier.exception.ResourceNotFoundException;
+import by.itech.lab.supplier.repository.ApplicationItemRepository;
 import by.itech.lab.supplier.repository.ItemRepository;
 import by.itech.lab.supplier.service.CategoryService;
 import by.itech.lab.supplier.service.ItemService;
@@ -27,7 +28,7 @@ public class ItemServiceImpl implements ItemService {
 
     private final ItemMapper itemMapper;
 
-    private final CategoryService categoryService;
+    private final ApplicationItemRepository applicationItemRepository;
 
     @Override
     public Page<ItemDto> findByLabel(final String label, final Pageable pageable) {
@@ -35,10 +36,9 @@ public class ItemServiceImpl implements ItemService {
                 .map(itemMapper::map);
     }
 
-    public Page<ItemDto> findAllByCategory(final String categoryName, final Pageable pageable) {
-        CategoryDto found = categoryService.findByCategory(categoryName);
-        return itemRepository.findAllByCategory(found.getId(), pageable)
-                .map(itemMapper::map);
+    public Page<ItemDto> findAllByCategory(final CategoryDto categoryDto, final Pageable pageable) {
+        return itemRepository.findAllByCategory(categoryDto.getId(), pageable)
+          .map(itemMapper::map);
     }
 
     @Override
@@ -59,7 +59,7 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public Page<ItemDto> findAll(Pageable pageable) {
+    public Page<ItemDto> findAll(final Pageable pageable) {
         return itemRepository.findAll(pageable)
                 .map(itemMapper::map);
     }
@@ -71,8 +71,13 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     @Transactional
-    public void delete(Long id) {
-        itemRepository.deleteById(id, LocalDate.now());
+    public void delete(final Long id) {
+        if (applicationItemRepository.getNumberOfItemUsages(id) == 0) {
+            itemRepository.deleteById(id, LocalDate.now());
+        } else {
+            throw new ResourceNotFoundException("Item can not be deleted, because it is already used in " +
+                    "application");
+        }
     }
 
     @Override
